# api/openapi.yaml
openapi: 3.0.3
info:
  title: Muasal API
  version: 0.1.0
  description: >-
    The API the web app uses, open to scripts through personal access tokens (FSD §14.3). Send
    `Authorization: Bearer msl_…`; a token acts with its owner's visibility and role, a read-only token only reads
    (and may ask), and each token takes 60 requests a minute, bursting to 120. Browser sessions use the sid cookie
    and must send the app's Origin on writes.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
paths:
...
  /projects/{key}/nodes:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string } }
    get:
      operationId: listNodes
      tags: [nodes]
      description: The tree as a flat list. Client-specific menus outside the caller's scope, and what is under them, are left out.
      parameters:
        - { name: archived, in: query, schema: { type: boolean }, description: Project admins only; also list archived nodes. }
      responses:
        "200":
          description: Live nodes; siblings in position order.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeList" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createNode
      tags: [nodes]
      description: Project admins only. The node goes last among its siblings.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeCreate" }
      responses:
        "201":
          description: The new node.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        default: { $ref: "#/components/responses/Problem" }
...
  /projects/{key}/nodes/recent:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string } }
    get:
      operationId: listRecentNodes
      tags: [nodes]
      description: >-
        The menus and modules of the caller's own latest tickets in the project, most recent first, at most 8, for
        "recently used first" in the menu picker (FSD §8.1). Only live nodes the caller sees.
      responses:
        "200":
          description: Node ids, most recent first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecentNodes" }
        default: { $ref: "#/components/responses/Problem" }
  /nodes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: getNode
      tags: [nodes]
      description: Every member who sees the node, archived or not (R-MR-4), with the path of its parents.
      responses:
        "200":
          description: The node and its path.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeDetail" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      operationId: updateNode
      tags: [nodes]
      description: Project admins only. Edits fields, and moves the node when `move` is present.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeUpdate" }
      responses:
        "200":
          description: The updated node.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteNode
      tags: [nodes]
      description: Project admins only. A node with sub-nodes answers 409 node_has_children.
      responses:
        "204": { description: Deleted. }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  responses:
    Problem:
      description: An error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
...
    NodeType:
      type: string
      enum: [module, menu]
    NodeClient:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    Node:
      type: object
      required: [id, parent_id, type, name, code, aliases, description, client_specific, clients, position, archived]
      properties:
        id: { type: integer, format: int64 }
        parent_id: { type: integer, format: int64, nullable: true }
        type: { $ref: "#/components/schemas/NodeType" }
        name: { type: string }
        code: { type: string, nullable: true, example: HR.ATT.OT_APPROVAL }
        aliases:
          type: array
          items: { type: string }
        description: { type: string }
        client_specific: { type: boolean }
        clients:
          type: array
          description: The node's clients within the caller's scope.
          items: { $ref: "#/components/schemas/NodeClient" }
        position: { type: integer, format: int32 }
        archived: { type: boolean }
    NodeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Node" }
    NodeCreate:
      type: object
      required: [type, name]
      properties:
        parent_id: { type: integer, format: int64 }
        type: { $ref: "#/components/schemas/NodeType" }
        name: { type: string, maxLength: 200 }
        code: { type: string, maxLength: 100 }
        aliases:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 5000 }
        client_specific: { type: boolean }
        client_ids:
          type: array
          items: { type: integer, format: int64 }
    NodeUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 200 }
        type: { $ref: "#/components/schemas/NodeType" }
        code: { type: string, maxLength: 100, description: An empty code clears it. }
        aliases:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 5000 }
        client_specific: { type: boolean, description: When present, client_ids replaces the node's clients. }
        client_ids:
          type: array
          items: { type: integer, format: int64 }
        move: { $ref: "#/components/schemas/NodeMove" }
        archived: { type: boolean, description: Archive (true) or restore (false); a linked node is archived, never deleted. }
    NodeMove:
      type: object
      properties:
        parent_id: { type: integer, format: int64, nullable: true, description: Omitted or null moves to the top level. }
        position: { type: integer, format: int32, minimum: 0, description: Index among the new siblings; omitted puts the node last. }
    NodeDetail:
      type: object
      required: [node, path, project_key]
      properties:
        node: { $ref: "#/components/schemas/Node" }
        path:
          type: array
          description: The node's parents, top first.
          items: { $ref: "#/components/schemas/Ref" }
        project_key: { type: string }
...
    Ref:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
    RecentNodes:
      type: object
      required: [node_ids]
      properties:
        node_ids:
          type: array
          items: { type: integer, format: int64 }